% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctfac.R
\name{correctfac}
\alias{correctfac}
\title{Calculate correction factor matrix from bulk data.}
\usage{
correctfac(type_bulk, nei)
}
\arguments{
\item{nei}{neighborhood size}

\item{matrix}{bulk matrix}
}
\value{
a matrix of bulk data correction factor
}
\description{
This function calculates the correction factor that borrows information from bulk data. If bulk data is available, it can be used to modify the prior setting in Bayesian hierarchical model.
}
\examples{
correctfac(m, nei=2) #calculate bulk data correction factor, m is the bulk matrix
}
