% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{hm}
\alias{hm}
\title{hm}
\usage{
hm(datvec, n, title = "Heatmap")
}
\arguments{
\item{datvec}{A vector of upper triangular mamtrix.}

\item{n}{Number of bins.}

\item{title}{The title of the heatmap.}
}
\value{
Heatmap of the matrix.
}
\description{
This function draws heatmap of HiC data so that we can visually compares the imputation results.
}
\examples{
.
data("K562_1_true")
hm(K562_1_true[,1], 61, title="Expected")
}
