% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrow.R
\name{matrow}
\alias{matrow}
\title{Calculate the position in the vector}
\usage{
matrow(x, y)
}
\arguments{
\item{x}{bin1}

\item{y}{bin2}
}
\value{
position in the vector
}
\description{
This function calculates the position of a pair in the vector when we transform the upper triangular matrix to a vector.
}
\examples{
matrow(2,3) #calculate the position of bin2 and bin3 in the vector.
}
