% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattovec.R
\name{mattovec}
\alias{mattovec}
\title{Converts the upper triangular matrix to a vector.}
\usage{
mattovec(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A vector of the upper triangular of the matrix.
}
\description{
This function converts the upper triangular part of a matrix to a vector
}
\examples{
mattovec(m)
}
