% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCImpute.R
\name{MCMCImpute}
\alias{MCMCImpute}
\title{Impute single cell HiC data}
\usage{
MCMCImpute(
  niter = 30000,
  burnin = 15000,
  single,
  bulk = bulk,
  startval = c(100, 100, 10, 8, 10, 0.1, 900, 0.2, 0, replicate(dim(single)[2], 8)),
  n,
  epsilon1 = 0.5,
  epsilon2 = 5,
  mc.cores = 1,
  cutoff = 0.5
)
}
\arguments{
\item{niter}{Number of MCMC iteration.}

\item{burnin}{Number of MCMC burnin.}

\item{single}{Single cell matrix with each column being the upper triangular matrix of a single cell.}

\item{bulk}{Bulk data. If NULL, bulk is set to be the sum of all single cells.}

\item{startval}{starting value of MCMC chain.}

\item{n}{Dimension of single cell matrix(original matrix is square and symmetrical).}

\item{epsilon1}{Range size of delta, default is 0.5.}

\item{epsilon2}{Range size of B, default is 5.}

\item{mc.cores}{Number of cores to be used in mclapply function, default is 1.}

\item{cutoff}{The threshold of \eqn{\pi_{ij}} that is used to define structural zeros.}
}
\value{
A list of posterior mean of probability, imputed data without defining SZ, and imputed data with SZ.
}
\description{
This function imputes the single cell data by borrowing information from beighborbood, similar cells, and bulk data.
}
\examples{
data("K562_T1_7k")
data("K562_bulk")
single=K562_T1_7k
T1_7k_res=MCMCImpute(niter=100000,burnin=5000,single=K562_T1_7k,bulk=K562_bulk,
startval=c(100,100,10,8,10,0.1,900,0.2,0,replicate(dim(single)[2],8)),n=61,mc.cores = 1,cutoff=0.5)
}
