% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neivar.R
\name{neivar}
\alias{neivar}
\title{Calculate neighborhood standard deviation.}
\usage{
neivar(single, nei, n)
}
\arguments{
\item{single}{matrix with each column being the upper triangular of each single cell}

\item{nei}{neighborhood size}

\item{n}{matrix size}
}
\value{
standard deviation matrix with each position being the standard deviation in the neighborhood among all single cells.
}
\description{
This function calculates the standard deviation of nonzero observations among single cells.
}
\examples{
neivar(m,nei=2) 

}
