% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{performance}
\usage{
performance(observed, expected, imputed)
}
\arguments{
\item{observed}{Observed observed cells matrix, with each column being the upper triangular of a observed cell.}

\item{expected}{Underline true count of the simulated data.}

\item{imputed}{Imputed matrix from MCMCImpute.}
}
\value{
A list of accuracy measurements with mean and standard error.
}
\description{
This function summarizes the accuracy of simulated data.
}
\examples{
data("K562_1_true")
options(digits = 2)
performance(observed=K562_T1_7k, expected=K562_1_true, imputed=T1_7k_imp)
}
