% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTDO80.R
\name{PTDO80}
\alias{PTDO80}
\title{PTDO80}
\usage{
PTDO80(observed, expected, result)
}
\arguments{
\item{observed}{Observed single cells matrix with each column being the upper triangular of a single cell.}

\item{expected}{Underline true counts from simulation.}

\item{result}{Result from MCMCImpute funtion.}
}
\value{
A vector of PTSZ and its SD when fixing PTDO to be 0.80, and the threshold used in that case.
}
\description{
This function calculates PTDO when fix PTDO=0.80.
}
\examples{
PTDO80(observed=K562_T1_7k, expected=K562_1_true, result=T1_7k_res)
}
