% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEVI.R
\name{SEVI}
\alias{SEVI}
\title{SEVI}
\usage{
SEVI(observed, expected, imputed)
}
\arguments{
\item{observed}{A vector of observed single cell.}

\item{expected}{A vector of expected single cell.}

\item{imputed}{A vector of imputed single cell.}
}
\value{
The scatterplot of expected versus imputed, with read dots being the observed zero pairs.
}
\description{
SEVI
}
\examples{
SEVI(observed=K562_T1_7k[,1], expected=K562_1_true[,1], imputed=T1_7k_imp[,1] )
}
