% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudat.R
\docType{data}
\name{simudat}
\alias{simudat}
\title{simudat}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1830 rows and 10 columns.
}
\usage{
data(simudat)
}
\description{
Simulated single cell from str1.
}
\examples{
data(simudat)
}
\keyword{datasets}
