% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudat_res.R
\docType{data}
\name{simudat_res}
\alias{simudat_res}
\title{simudat_res}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(simudat_res)
}
\description{
result of MCMCImput on simudat.
}
\examples{
data(simudat_res)
}
\keyword{datasets}
