% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudat_true.R
\docType{data}
\name{simudat_true}
\alias{simudat_true}
\title{simudat_true}
\format{
An object of class \code{data.frame} with 1830 rows and 10 columns.
}
\usage{
data(simudat_true)
}
\description{
Underline truecount of simudat.
}
\examples{
data(simudat_true)
}
\keyword{datasets}
