% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOVI.R
\name{SOVI}
\alias{SOVI}
\title{SOVI}
\usage{
SOVI(observed, imputed)
}
\arguments{
\item{observed}{A vector of observed single cell.}

\item{imputed}{A vector of imputed single cell.}
}
\value{
The scatterplot of observed versus imputed.
}
\description{
Scatterplot of observed versus imputed at nonzeros.
}
\examples{
data("GSE117874_imp")
SOVI(observed=GSE117874_chr1_wo_diag[,1], imputed = GSE117874_imp[,1])
}
