% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str1.R
\docType{data}
\name{str1}
\alias{str1}
\title{str1}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 61 rows and 3 columns.
}
\usage{
data(str1)
}
\description{
A 3D coordinates of chromosome. 
The structure was generated from SIMBA package.
}
\examples{
data(str1)
}
\keyword{datasets}
